Add TUs

+++ app/src/main/java/fr/delcey/mvvm_clean_archi/view/model/PropertyUiModel.kt
@@ -1,7 +1,7 @@
 package fr.delcey.mvvm_clean_archi.view.model
 
 data class PropertyUiModel(
-    val id : Int, // Shared between Property and Address
+    val id : Long, // Shared between Property and Address
     val type : String, // Coming from Propery
     val mainAddress : String? = null // Coming from Address
 )
